layers: Move Pipeline vertex code to LastBound
